% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignFastqBwa.R
\name{alignFastqBwa}
\alias{alignFastqBwa}
\title{alignFastqBwa}
\usage{
alignFastqBwa(input_file, ref, out_path, threads)
}
\arguments{
\item{input_file}{Fastq file to align (in paired use R1)}

\item{ref}{Path where the reference genome is save..}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}
}
\description{
This function aligns fq files using a defined reference genome and bwa being possible to
use paired or single end data
}
\examples{
\dontrun{
input_file <- 'raw/sample_R1.fa'
ref <- 'ref/hg38.fa'
out_path <- 'rst'
threads <- 2

alignFastqBwa(input_file, ref, out_path, threads, removeOpt)
}

}
