% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovarAnnotation.R
\name{annovarAnnotation}
\alias{annovarAnnotation}
\title{annovarAnnotation}
\usage{
annovarAnnotation(
  vcf,
  out_vcf,
  threads,
  annovar = "/imppc/labs/lplab/share/bin/annovar",
  humandb = "/imppc/labs/lplab/share/bin/annovar/humandb",
  perl = "perl"
)
}
\arguments{
\item{vcf}{Vcf file to annotate}

\item{out_vcf}{Path where to save the annotated vcf.}

\item{threads}{Number of threads to use in the analysis.}

\item{annovar}{Path to the annovar binary folder.}

\item{humandb}{Path to the annovar huban database.}

\item{perl}{Path of perl executable.}
}
\description{
R parser for Annovar
}
\examples{
\dontrun{
annovarAnnotation(vcf = 'NET-11_out_strelka2_indels_postCalling.vcf',
                  out_vcf = 'NET-11_annotated.vcf',
                  threads = 2)
}

}
