% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamCoverage.R
\name{bamCoverage}
\alias{bamCoverage}
\title{bamCoverage}
\usage{
bamCoverage(
  deep_tools = "/imppc/labs/lplab/share/bin/anaconda3/bin/activate",
  bam,
  cores = 2,
  out_path
)
}
\arguments{
\item{deep_tools}{Path to the activate file of the deeptools virtual environment}

\item{bam}{Bam file to generate the bam counts}

\item{cores}{Number of cores used in the analysis}

\item{out_path}{Path where to save the output bigWig. It will be saved with the same base
name as bam and ".bw" extension.}
}
\description{
Parser of DeepTools BamCounter for R
}
\examples{
out_path <- '/counts
bam <- '/raw/sample.bam

bamCoverage(bam, out_path)
}
