% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwaAlignment.R
\name{bamTofastqPicard}
\alias{bamTofastqPicard}
\title{bamTofastqPicard}
\usage{
bamTofastqPicard(input_file, ref, out_path, threads, gatk4, sambamba)
}
\arguments{
\item{input_file}{Fastq file to align. If paried-end type, `input_file` have to contain mate 1s and different pairs should
be named "_R1" or "_R2". Allowed formats: fastq.gz, fq.gz, fastq, fq or bam.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{gatk4}{Path of GATK4 binary.}

\item{sambamba}{Path of sambamba binary.}
}
\description{
This function converts BAM files to FASTQ files using GATK4 SamToFastq. Type of BAM file (paried or single) is detected
automatically.
}
\examples{
\dontrun{
bamTofastqPicard(input_file = 'raw/sample.bam',
             ref = 'ref/hg38.fa',
             out_path = 'rst',
             threads = '2')
}
}
