% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftoolsReheader.R
\name{bcftoolsReheader}
\alias{bcftoolsReheader}
\title{bcftoolsReheader}
\usage{
bcftoolsReheader(
  bcftools,
  vcf,
  vcf_new_names = c("NORMAL", "TUMOR"),
  reheadered_vcf
)
}
\arguments{
\item{vcf}{Vcf to reheader.}

\item{vcf_new_names}{Vector with the new sample names to define in the vcf header.}

\item{reheadered_vcf}{Path where to save the reheared vcf.}
}
\description{
Parser for bcftools reheader
}
\examples{
\dontrun{
bcftoolsReheader(bcftools = 'bcftools', vcf = 'sample.vcf', vcf_new_names = c('NORMAL', 'TUMOR'), reheadered_vcf = 'sample_rh.vcf')
}
}
