% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwaAlignment.R
\name{bwaAlignment}
\alias{bwaAlignment}
\title{bwaAlignment}
\usage{
bwaAlignment(
  fastq,
  ref,
  out_path,
  threads,
  sambamba = "sambamba",
  bwa = "bwa",
  samblaster = "samblaster",
  samtools = "samtools"
)
}
\arguments{
\item{fastq}{Fastq file to carry the analysis. If paried-end type, `input_file` have to contain mate 1s and different pairs should
be named "_R1" or "_R2". Allowed formats: fastq.gz, fq.gz, fastq, fq or bam.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{sambamba}{Path of sambamba binary}

\item{bwa}{Path of bwa binary.}

\item{samblaster}{Path of samblaster binary.}

\item{samtools}{Path of samtools binary.}
}
\description{
Aligns FASTQ files using bwa mem and a defined reference genome. It alignes fastq paired or single end data. The
BAM to FASTQ conversion can be carried out using bamTofastqPicard function.
}
\examples{
\dontrun{
bwaAlignment(fastq = 'raw/sample_R1.fa',
             ref = 'ref/hg38.fa',
             out_path = 'rst',
             threads = 2)
}
}
