% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets}
\alias{facets}
\title{facets}
\usage{
facets(normal, tumor, outName, outPath)
}
\arguments{
\item{normal}{Normal bam file}

\item{tumor}{Tumour bam file}

\item{outName}{Name for the output file for the paired tumor normal analysis}

\item{outPath}{Path where the output of the analysis will be saved.}
}
\description{
This function carries out allele-specific copy number and clonal heterogeneity analysis tool for high-throughput DNA sequencing
}
\examples{
\dontrun{
normal <- 'raw/sample12_norm.bam'
tumor <- 'raw/sample12_tum.bam'
outName <- 'sample12'
outPath <- 'rst/facets'

facets(inputFile, threads, outPath)
}

}
