% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPass.R
\name{filterPass}
\alias{filterPass}
\title{filterPass}
\usage{
filterPass(bcftools = "bcftools", vcf, pass_vcf)
}
\arguments{
\item{vcf}{Vcf to filter}

\item{out_path}{Path where to save the output vcf}
}
\description{
Bcftools filter parser for only keeping PASS variants of a VCF.
}
\examples{
\dontrun{
filterPass(vcf = 'raw/sample_tumor.vcf',
            out_path = 'pass_vcf')
}

}
