% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{fpfilterTool}
\alias{fpfilterTool}
\title{fpfilterTool}
\usage{
fpfilterTool(vcf, tumor_file, sample_name, ref, fpfilter, perl, out_path,
  bam_readcount, tumor_vcf_id = "TUMOR")
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{perl}{Path of perl executable.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{tumor_vcf_id}{Id of the tumor sample in vcf. By default 'TUMOR'.}
}
\description{
False postive filter parser for label low quality reads
}
\examples{
\dontrun{
fpfilterTool(vcf = 'vcf/sample.vcf',
             tumor_file = 'raw/sample_tumor.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst',
               perl = 'perl',
               fpfilter = 'fpfilter-tool/fpfilter.pl',
               bam_readcount = 'bam_readcount')
}
}
