% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2R.R
\name{hisat2R}
\alias{hisat2R}
\title{hisat2R}
\usage{
hisat2R(threads, type, index, fastq_file, out_path)
}
\arguments{
\item{type}{"single" or "paired" format}

\item{index}{HISAT2 reference genome index}

\item{fastq_file}{Fastq to demultiplex.
In paired-end file used should be R1 and following the format "file_R1.*"}

\item{out_path}{Path where to save the demultiplex output}
}
\description{
Parser of HISAT2 for R
}
\examples{
threads <- 2
index <- '/imppc/labs/lplab/share/marc/refgen/hg38/HISAT2/hg38.tar.gz'
fastq_file <- '/imppc/labs/lplab/share/Insulinoma/raw/NET20_INS_RNA_1.fq.gz'
out_path <- '/imppc/labs/lplab/share/Insulinoma/raw/proves'
type <- 'paired'

hisat2R(threads, type, index, fastq_file, out_path)
}
