% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{insertSizePindel}
\alias{insertSizePindel}
\title{insertSizePindel}
\usage{
insertSizePindel(
  out_path,
  sambamba,
  samtools,
  bam_file,
  size_subsample = 1e+06
)
}
\arguments{
\item{out_path}{Path where the output of the analysis will be saved.}

\item{sambamba}{Path of sambamba binary}

\item{samtools}{Path of samtools binary.}

\item{bam_file}{BAM file where to calculate the insert size.}

\item{size_subsample}{Number of reads in the BAM file used for the insert size calculation.}
}
\description{
Calculates the mean insert size and returns the name of a bam-file,
the expected average insert size, and a label to indicate the identity of the sample.
}
\examples{
\dontrun{
insertSizePindel(bam_file = 'raw/sample_tumor.bam',
               samtools = '/bin/samtools',
               sambamba = '/bin/sambamba')
}

}
