% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manta.R
\name{manta}
\alias{manta}
\title{manta}
\usage{
manta(normal, tumor, ref, outPath, cores, outName)
}
\arguments{
\item{normal}{Normal bam file.}

\item{tumor}{Tumour bam file.}

\item{ref}{Reference genome.}

\item{outPath}{Path where the output of the analysis will be saved.}

\item{cores}{Number of cores to use.}

\item{outName}{Name for the output file for the paired tumor normal analysis}
}
\description{
SVs calling of paired tumor-normal wgs using manta
}
\examples{
\dontrun{

normal <- 'raw/sample1_BL.bam'
tumor <- 'raw/sample1_TI.bam'
ref <- '/imppc/labs/lplab/share/marc/refgen/hg38/hg38.fa'
outPath <- 'dir/vcf'
cores <- 4
outName <- 'sample1'
manta(normal = normal,
      tumor = tumor,
      ref = ref,
      outPath = outPath,
      cores = cores,
      outName = outName)

}
}
