% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{mantaCalling}
\alias{mantaCalling}
\title{mantaCalling}
\usage{
mantaCalling(tumor_file, normal_file, sample_name, ref, out_path, manta,
  threads)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{manta}{Path of manta binary.}

\item{threads}{Number of threads to use in the analysis.}
}
\description{
Tumor-normal pair somatic variant calling using manta.
}
\examples{
\dontrun{
mantaCalling(tumor_file = 'raw/sample_tumor.bam',
               normal_file = 'raw/sample_normal.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst',
               threads = 2,
               manta = '/bin/manta-1.4.0')
}

}
