% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdepth.R
\name{mosdepth}
\alias{mosdepth}
\title{mosdepth}
\usage{
mosdepth(inputFile, threads, outPath)
}
\arguments{
\item{inputFile}{Path of the fastq file to apply depth calculation.}

\item{threads}{Specify number of threads to use .}

\item{outPath}{Path where the output of the analysis will be saved.}
}
\description{
Fast BAM/CRAM depth calculation for WGS, exome, or targeted sequencing
}
\examples{
\dontrun{
inputFile <- 'raw/sample.bam'
ref <- 'ref/hg38.fa'
outPath <- 'rst/coverage'
threads <- 2

mosdepth(inputFile, threads, outPath)
}

}
