% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{museCalling}
\alias{museCalling}
\title{museCalling}
\usage{
museCalling(tumor_file, normal_file, sample_name, ref, out_path, muse,
  af_only_gnomad, bcftools)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{muse}{Path of MuSE binary.}

\item{af_only_gnomad}{Genome aggregation database used as a germline resource. Have to be base on the same reference genome
as `ref`. \href{https://gnomad.broadinstitute.org/downloads}{gnomAD}}
}
\description{
Tumor-normal pair somatic variant calling using MuSE
}
\examples{
\dontrun{
museCalling(tumor_file = 'raw/sample_tumor.bam',
            normal_file = 'raw/sample_normal.bam',
            sample_name = 'sample'
            ref = 'ref/hg38.fa',
            out_path = 'rst',
            muse = '/bin/MuSE',
            af_only_gnomad = 'hg38/af-only-gnomad.hg38.vcf.gz')
}

}
