% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncodriveclustl.R
\name{oncodriveclustl}
\alias{oncodriveclustl}
\title{oncodriveclustl}
\usage{
oncodriveclustl(
  input,
  regions,
  output_dir,
  genome = "hg38",
  cores = 2,
  kmer = 5,
  sigcalc = "region_normalized",
  sw = 11,
  simw = 31,
  cw = 11
)
}
\arguments{
\item{input}{File containing mutations}

\item{regions}{GZIP compressed file with the genomic regions to analyze}

\item{output_dir}{Output directory to be created}

\item{genome}{Genome to use. Default hg38.}

\item{cores}{Cores to use in the analysis.}

\item{kmer}{K-mer nucleotide context. Default 5.}

\item{sigcalc}{Signature calculation: mutation frequencies (frequencies) or k-mer mutation counts
normalized by k-mer region counts (region_normalized). Default region_normalized}

\item{sw}{Smoothing window. Default is 11}

\item{simw}{Simulation window. Default is 31}

\item{cw}{Cluster window. Default is 11}
}
\description{
Parser for oncodriveclustl
}
\examples{
\dontrun{
oncodriveclustl(input = 'all_ins.tvs', regions = 'regions.regions', output_dir = 'rst/', genome = 'hg38', cores = 2)
}

}
