% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncodrivefml.R
\name{oncodrivefml}
\alias{oncodrivefml}
\title{oncodrivefml}
\usage{
oncodrivefml(input, elements, output_dir, sequencing = "wgs", configuration)
}
\arguments{
\item{input}{File containing mutations}

\item{elements}{GZIP compressed file with the genomic elements to analyze}

\item{output_dir}{Output directory to be created}

\item{sequencing}{Type of sequencing. Allowd 'wgs', 'wes'
'targeted'. By default='wgs'.}

\item{configuration}{Path to configuration file.}
}
\description{
Parser for oncodrivefml.
}
\examples{
\dontrun{
oncodrivefml(input = 'all_ins.tvs', elements = 'elements.elements', output_dir = 'rst/')
}

}
