% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{pindelCalling}
\alias{pindelCalling}
\title{pindelCalling}
\usage{
pindelCalling(
  tumor_file,
  normal_file,
  sample_name,
  ref,
  out_path,
  pindel,
  sambamba,
  samtools,
  threads,
  perl,
  gatk4,
  centromeres_telomeres,
  bcftools
)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{pindel}{Path of Pindel binary.}

\item{sambamba}{Path of sambamba binary}

\item{samtools}{Path of samtools binary.}

\item{threads}{Number of threads to use in the analysis.}

\item{perl}{Path of perl executable.}

\item{gatk4}{Path of GATK4 binary.}

\item{centromeres_telomeres}{Bed file with the centromers and/or telomeres base on the same reference genome as `ref`.}
}
\description{
Tumor-normal pair somatic variant calling using Pindel
}
\examples{
\dontrun{
pindelCalling(tumor_file = 'raw/sample_tumor.bam',
               normal_file = 'raw/sample_normal.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst')
}

}
