% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAlignment.R
\name{postAlignment}
\alias{postAlignment}
\title{postAlignment}
\usage{
postAlignment(input_file, ref, out_path, threads,
  gatk4 = "/imppc/labs/lplab/share/bin/gatk-4.1.3.0/gatk",
  known_sites = c("/imppc/labs/lplab/share/marc/refgen/hg38/Mills_and_1000G_gold_standard.indels.hg38.vcf"))
}
\arguments{
\item{input_file}{Fastq file to align. If paried-end type, `input_file` have to contain mate 1s and different pairs should
be named "_R1" or "_R2". Allowed formats: fastq.gz, fq.gz, fastq, fq or bam.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{gatk4}{Path of GATK4 binary}

\item{known_sites}{Vector object with the GATK databases of known polymorphic sites
used to exclude regions around known polymorphisms from analysis. They have to be based on the same reference genome as `ref`.
#' @examples
\dontrun{
postAlignment(input_file = 'raw/sample.bam',
             ref = ref/hg38.fa',
             out_path = 'rst',
             threads = '2')
}}
}
\description{
Local realignment of insertions and deletions is performed using GATK IndelRealigner and
a base quality score recalibration (BQSR) step performed using GATK BaseRecalibrator.
Local realigment of INDELs locates regions that contain misalignments across BAM files,
which can often be caused by insertion-deletion (indel) mutations with respect to the reference genome.
BQSR adjusts base quality scores based on detectable and systematic errors.
}
