% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postCalling.R
\name{postCalling}
\alias{postCalling}
\title{postCalling}
\usage{
postCalling(
  pass_vcf,
  gatk4 = "/imppc/labs/lplab/share/bin/gatk-4.1.3.0/gatk",
  ref,
  bcftools = "bcftools",
  out_path
)
}
\arguments{
\item{pass_vcf}{Vcf to apply post-calling processing. It should be descompressed and *_PASS.vcf format.}

\item{gatk4}{Path of GATK4 binary.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}
}
\description{
Vcf sorting, normalization and reheader of the vcf generated unsing `variantCalling` for posterior annotation. Input file have to
present the *_PASS.vcf format and being uncompressed.
}
\examples{
\dontrun{
postCalling(pass_vcf = 'raw/sample_tumor.bam',
            gatk4 = 'bin/gatk-4.1.3.0/gatk',
            ref = 'ref/hg38.fa',
            bcftools = 'bcftools')
}

}
