% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{radiaCalling}
\alias{radiaCalling}
\title{radia}
\usage{
radiaCalling(
  tumor_file,
  normal_file,
  sample_name,
  ref,
  out_path,
  radia_path,
  python_radia,
  samtools,
  threads,
  bcftools
)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{python_radia}{Path to the python binary with all the RADIA \href{https://github.com/aradenbaugh/radia}{prerequisites}.}

\item{samtools}{Path of samtools binary.}

\item{threads}{Number of threads to use in the analysis.}
}
\description{
Tumor-normal pair somatic variant calling using RADIA
}
\examples{
\dontrun{
radiaCalling(tumor_file = 'raw/sample_tumor.bam',
               normal_file = 'raw/sample_normal.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst',
               radia_path = '/bin/radia/scripts',
               python_radia = 'venv/radia/bin/python',
               samtools = '/bin/samtools')
}

}
