% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realignBamBwa.R
\name{realignBamBwa}
\alias{realignBamBwa}
\title{realignBamBwa}
\usage{
realignBamBwa(inputFile, ref, outPath, threads, removeOpt)
}
\arguments{
\item{inputFile}{Path of the fastq file to align.}

\item{ref}{Path where the reference genome is save.}

\item{outPath}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{removeOpt}{Remove old bam option (yes or no)}
}
\description{
This function realigns bam files using a defined reference genome and bwa being possible to
use paired or single end data (type will be detected automatically)
}
\examples{
\dontrun{
inputFile <- 'raw/sample.bam'
ref <- 'ref/hg38.fa'
outPath <- 'rst'
threads <- 2
removeOpt <- 'yes'

realignBamBwa(inputFile, ref, outPath, threads, removeOpt)
}

}
