% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reheaderBam.R
\name{reheaderBam}
\alias{reheaderBam}
\title{reheaderBam}
\usage{
reheaderBam(bam, out_path, samtools = "samtools")
}
\description{
Samtools R parser for reheader bams with the bam file name. 
Output file will have the same name as input BAM, so define a different output folder from the input folder.
}
