% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameSampleBam.R
\name{renameSampleBam}
\alias{renameSampleBam}
\title{renameSampleBam}
\usage{
renameSampleBam(bam, samtools = "samtools", new_sample_name, out_path)
}
\arguments{
\item{samtools}{Path of samtools binary.}

\item{new_sample_name}{New sample name for the BAM file}

\item{out_path}{Path where the output of the analysis will be saved.}
}
\description{
Rename sample name of a BAM using samtools.
Input folder and output folder should be different for avoding overwrite problems.
}
\examples{
\dontrun{
renameSampleBam(bam = 'sample.bam',
                samtools = 'samtools',
                new_sample_name = 'new_sm',
                out_path = 'bam/renamed')
}

}
