% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{samtoolsIndex}
\alias{samtoolsIndex}
\title{sambambaIndex}
\usage{
samtoolsIndex(bam, threads, samtools)
}
\arguments{
\item{threads}{Number of threads to use in the analysis.}

\item{samtools}{Path of samtools binary.}
}
\description{
Generate BAM index if it is not present using sambamba
}
\examples{
\dontrun{
sambambaIndex(bam = 'raw/sample_tumor.bam',
            threads = 2,
            sambamba = '/bin/sambamba')
}

}
