% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{somaticsniperCalling}
\alias{somaticsniperCalling}
\title{somaticsniperCalling}
\usage{
somaticsniperCalling(
  tumor_file,
  normal_file,
  sample_name,
  ref,
  out_path,
  somatic_sniper,
  fpfilter,
  perl,
  bcftools
)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{somatic_sniper}{Path of SomaticSniper binary.}

\item{perl}{Path of perl executable.}
}
\description{
Tumor-normal pair somatic variant calling using SomaticSniper
}
\examples{
\dontrun{
somaticsniperCalling(tumor_file = 'raw/sample_tumor.bam',
               normal_file = 'raw/sample_normal.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst',
               somatic_sniper = '/bin/bam-somaticsniper')
}

}
