% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticWgsAnalysis.R
\name{somaticWgsAnalysis}
\alias{somaticWgsAnalysis}
\title{somaticWgsAnalysis}
\usage{
somaticWgsAnalysis()
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{threads}{Number of threads to use in the analysis.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{muse}{Path of MuSE binary.}

\item{gatk4}{Path of GATK4 binary.}

\item{samtools_mpileup}{Path of samtools mpileup binary.}

\item{af_only_gnomad}{Genome aggregation database used as a germline resource. Have to be base on the same reference genome
as `ref`. \href{https://gnomad.broadinstitute.org/downloads}{gnomAD}}

\item{somatic_sniper}{Path of SomaticSniper binary.}

\item{bwa}{Path of bwa binary.}

\item{samblaster}{Path of samblaster binary.}

\item{samtools}{Path of samtools binary.}

\item{sambamba}{Path of sambamba binary}

\item{indel_candidates}{For the somatic workflow, the best-practice recommendation is to run the Manta SV and indel caller on the same set of samples first,
then supply Manta's candidate indels as input to Strelka. Defined sample name have to be the same in both, Strelka2 and manta for
a correct detection of the indel candidates file.}

\item{centromeres_telomeres}{Bed file with the centromers and/or telomeres base on the same reference genome as `ref`.}

\item{varscan}{Path of Varscan2 binary.}

\item{manta}{Path of manta binary.}

\item{strelka2}{Path of strelka2 binary.}

\item{pindel}{Path of Pindel binary.}

\item{perl}{Path of perl executable.}

\item{fastq}{Fastq file to carry the analysis. If paried-end type, `input_file` have to contain mate 1s and different pairs should
be named "_R1" or "_R2". Allowed formats: fastq.gz, fq.gz, fastq, fq or bam.}

\item{python_radia}{Path to the python binary with all the RADIA \href{https://github.com/aradenbaugh/radia}{prerequisites}.}

\item{tumor_vcf_id}{Id of the tumor sample in vcf. By default 'TUMOR'.}

\item{bam}{Bam file to carry the analysis.}
}
\description{
somaticWgsAnalysis pipeline identifies somatic variants within hole genome sequencing (WGS) data.
The first pipeline starts with a reference alignment step followed by co-cleaning to increase the alignment quality.
Six different variant calling pipelines are then implemented separately to identify somatic mutations.
\itemize{
  \item \href{https://bioinformatics.mdanderson.org/public-software/muse/}{MuSE}
  \item \href{https://gatk.broadinstitute.org/hc/en-us/articles/360035531132--How-to-Call-somatic-mutations-using-GATK4-Mutect2}{MuTect2}
  \item \href{http://varscan.sourceforge.net/}{VarScan2}
  \item \href{https://github.com/genome/somatic-sniper}{SomaticSniper}
  \item \href{http://gmt.genome.wustl.edu/packages/pindel/}{Pindel}
  \item \href{https://github.com/Illumina/strelka}{Strelka2}
  \item \href{https://github.com/Illumina/manta}{Manta}
}

Somatic-caller-identified variants are then annotated. Annotated VCF are converted into MAF file finally.
}
