% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka2.R
\docType{data}
\name{strelka}
\alias{strelka}
\title{strelka2}
\format{
An object of class \code{character} of length 1.
}
\usage{
strelka
}
\arguments{
\item{normal}{Normal bam file.}

\item{tumor}{Tumour bam file.}

\item{ref}{Reference genome.}

\item{outPath}{Path where the output of the analysis will be saved.}

\item{cores}{Number of cores to use.}

\item{outName}{Name for the output file for the paired tumor normal analysis}

\item{indelCandidates}{For the somatic workflow, the best-practice recommendation is to run the Manta SV and indel caller on the same set of samples first, then supply Manta's candidate indels as input to Strelka}
}
\description{
Snvs and indels calling of paired tumor-normal wgs using manta
}
\examples{
\dontrun{

normal <- 'raw/sample1_BL.bam'
tumor <- 'raw/sample1_TI.bam'
ref <- '/imppc/labs/lplab/share/marc/refgen/hg38/hg38.fa'
outPath <- 'dir/vcf'
cores <- 4
outName <- 'sample1'
indelCandidates <- /results/variants/candidateSmallIndels.vcf.gz
strelka2(normal = normal,
      tumor = tumor,
      ref = ref,
      outPath = outPath,
      cores = cores,
      outName = outName)

}
}
\keyword{datasets}
