% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantaStrelka2Calling.R
\name{strelka2Calling}
\alias{strelka2Calling}
\title{strelka2Calling}
\usage{
strelka2Calling(
  tumor_file,
  normal_file,
  sample_name,
  ref,
  out_path,
  threads,
  indel_candidates,
  strelka2,
  bcftools
)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{strelka2}{Path of strelka2 binary.}
}
\description{
Tumor-normal pair somatic variant calling using strelka2
}
\examples{
\dontrun{
strelka2Calling(tumor_file = 'raw/sample_tumor.bam',
               normal_file = 'raw/sample_normal.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst',
               threads = 2,
               indel_candidates = 'manta/results/variants/candidateSmallIndels.vcf.gz',
               strelka2 = '/bin/strelka-2.9.3')
}

}
