% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantAnnotation.R
\name{variantAnnotation}
\alias{variantAnnotation}
\title{variantAnnotation}
\usage{
variantAnnotation(vcf, vep, cache_dir, ref, out_path, sample_name)
}
\arguments{
\item{vcf}{File}

\item{vep}{File}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{cache_file}{Nan}
}
\description{
Raw VCF files are annotated using Variant Effect Predictor (\href{https://www.ensembl.org/info/docs/tools/vep/index.html}{VEP})
The VEP uses the coordinates and alleles in the VCF file to infer biological context for each variant including the location
of each mutation, its biological consequence (frameshift/ silent mutation), and the affected genes.
The following databases are used for VCF annotation:

\itemize{
  \item \href{https://www.ensembl.org/index.html}{Ensembl database}: Protein-coding and non-coding genes, splice variants,
  cDNA and protein sequences, non-coding RNAs, among others annotations.
  \item {https://www.gencodegenes.org/}{GENCODE}: human and mouse high accuracy based on biological evidence annotations.
  \item {https://www.ncbi.nlm.nih.gov/refseq/}{RefSeq}: A comprehensive, integrated, non-redundant,
  well-annotated set of reference sequences including genomic, transcript, and protein.
  \item {http://genetics.bwh.harvard.edu/pph2/}{PolyPhen}: Polymorphism Phenotyping v2 is a tool which predicts possible impact of an amino acid substitution
  on the structure and function of a human protein using straightforward physical and comparative considerations.
  \item {https://sift.bii.a-star.edu.sg/}{SIFT}:	SIFT predicts whether an amino acid substitution affects protein function
  based on sequence homology and the physical properties of amino acids. SIFT can be applied to naturally occurring
  nonsynonymous polymorphisms and laboratory-induced missense mutations.
  \item {https://www.ncbi.nlm.nih.gov/snp/}{dbSNP}: dbSNP contains human single nucleotide variations, microsatellites,
  and small-scale insertions and deletions along with publication, population frequency, molecular consequence,
  and genomic and RefSeq mapping information for both common variations and clinical mutations.
  \item {https://cancer.sanger.ac.uk/cosmic}{COSMIC}: the Catalogue Of Somatic Mutations In Cancer, is the world's largest
  and most comprehensive resource for exploring the impact of somatic mutations in human cancer.
  \item {http://www.hgmd.cf.ac.uk/ac/index.php}{HGMD-PUBLIC}: The Human Gene Mutation Database (HGMD®) represents an attempt to
  collate all known (published) gene lesions responsible for human inherited disease,
  \item {https://www.ncbi.nlm.nih.gov/clinvar/}{ClinVar}: reports of the relationships among human variations and phenotypes,
  with supporting evidence.
  \item \href{https://www.internationalgenome.org/home}{1000 Genomes}: database with most genetic variants with frequencies of
  at least 1% in the populations studied.
  \item NHLBI-ESP
  \item \href{https://gnomad.broadinstitute.org/}{gnomAD}: The Genome Aggregation Database (gnomAD) is a resource developed by
  an international coalition of investigators,with the goal of aggregating and harmonizing both exome and genome sequencing
  data from a wide variety of large-scale sequencing projects, and making summary data available for the wider scientific community.
}
}
\examples{
\dontrun{

}

}
