% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{variantCalling}
\alias{variantCalling}
\title{variantCalling}
\usage{
variantCalling(tumor_file, normal_file, sample_name, ref, out_path,
  threads = 1, gatk4 = "/imppc/labs/lplab/share/bin/gatk-4.1.3.0/gatk",
  muse = "/imppc/labs/lplab/share/marc/repos/MuSE/MuSE", bwa = "bwa",
  samblaster = "samblaster", sambamba = "sambamba",
  samtools = "samtools",
  somatic_sniper = "/imppc/labs/lplab/share/marc/repos/somatic-sniper/build/bin/bam-somaticsniper",
  varscan2 = "/imppc/labs/lplab/share/marc/repos/varscan/VarScan.v2.4.4.jar",
  manta = "/software/debian-8/bio/manta-1.4.0",
  strelka2 = "/soft/bio/strelka-2.9.3",
  af_only_gnomad = "/imppc/labs/lplab/share/marc/refgen/hg38/af-only-gnomad.hg38.vcf.gz",
  pindel = "/soft/bio/pindel/pindel",
  centromeres_telomeres = "/imppc/labs/lplab/share/marc/refgen/hg38/centromeres.bed",
  perl = "perl",
  python_radia = "/imppc/labs/lplab/share/marc/repos/miniconda/bin/python2.7",
  radia_path = "/imppc/labs/lplab/share/marc/repos/radia/scripts",
  bcftools = "bcftools", tumor_vcf_id = "TUMOR",
  fpfilter = "/imppc/labs/lplab/share/marc/repos/fpfilter-tool/fpfilter.pl",
  bam_readcount = "bam-readcount")
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{gatk4}{Path of GATK4 binary.}

\item{muse}{Path of MuSE binary.}

\item{bwa}{Path of bwa binary.}

\item{samblaster}{Path of samblaster binary.}

\item{sambamba}{Path of sambamba binary.}

\item{samtools}{Path of samtools binary.}

\item{somatic_sniper}{Path of SomaticSniper binary.}

\item{manta}{Path of manta binary.}

\item{strelka2}{Path of strelka2 binary.}

\item{af_only_gnomad}{Genome aggregation database used as a germline resource. Have to be base on the same reference genome
as `ref`. \href{https://gnomad.broadinstitute.org/downloads}{gnomAD}}

\item{pindel}{Path of Pindel binary.}

\item{centromeres_telomeres}{Bed file with the centromers and/or telomeres base on the same reference genome as `ref`.}

\item{perl}{Path of perl executable.}

\item{python_radia}{Path to the python binary with all the RADIA \href{https://github.com/aradenbaugh/radia}{prerequisites}.}

\item{tumor_vcf_id}{Id of the tumor sample in vcf. By default 'TUMOR'.}
}
\description{
Aligned and co-cleaned BAM files are processed as tumor-normal pairs.
Variant calling is performed using six separate pipelines:

\itemize{
  \item \href{https://bioinformatics.mdanderson.org/public-software/muse/}{MuSE}
  \item \href{https://gatk.broadinstitute.org/hc/en-us/articles/360035531132--How-to-Call-somatic-mutations-using-GATK4-Mutect2}{MuTect2}
  \item \href{http://varscan.sourceforge.net/}{VarScan2}
  \item \href{https://github.com/genome/somatic-sniper}{SomaticSniper}
  \item \href{http://gmt.genome.wustl.edu/packages/pindel/}{Pindel}
  \item \href{https://github.com/Illumina/strelka}{Strelka2}
  \item \href{https://github.com/Illumina/manta}{Manta}
  }
}
\examples{
\dontrun{
variantCalling(tumor_file = 'NET-10_TI_mkdup_sub_0005.bam',
               normal_file = 'NET-10_BL_mkdup_sub_0005.bam',
               sample_name = 'NET-10',
               ref = 'hg38.fa',
               out_path = 'processed/vcf',
               threads = 2)
}
}
