% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantCalling.R
\name{varscan2Calling}
\alias{varscan2Calling}
\title{varscan2Calling}
\usage{
varscan2Calling(
  tumor_file,
  normal_file,
  sample_name,
  ref,
  out_path,
  varscan2,
  samtools,
  fpfilter,
  perl,
  bcftools
)
}
\arguments{
\item{tumor_file}{Tumor bam to file to perform the variant calling.}

\item{normal_file}{Normal bam to file to perform the variant calling.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{samtools}{Path of samtools binary.}

\item{perl}{Path of perl executable.}
}
\description{
Tumor-normal pair somatic variant calling using VarScan2.
}
\examples{
\dontrun{
varscan2Calling(tumor_file = 'raw/sample_tumor.bam',
               normal_file = 'raw/sample_normal.bam',
               sample_name = 'sample'
               ref = 'ref/hg38.fa',
               out_path = 'rst',
               varscan2 = '/bin/VarScan.v2.4.4.jar',
               threads = 2,
               samtools = '/bin/samtools')
}

}
