% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsSomaticAnalysis.R
\name{wgsSomaticPreprocessing}
\alias{wgsSomaticPreprocessing}
\title{wgsSomaticPreprocessing}
\usage{
wgsSomaticPreprocessing(
  inputFile,
  ref,
  outPathBam,
  threads,
  removeOpt = "no",
  rstPath
)
}
\arguments{
\item{inputFile}{Fastq or bam file to analyse}

\item{ref}{Reference genome}

\item{outPathBam}{Path where to save the output bam}

\item{threads}{Number of threads to use in the analysis}

\item{removeOpt}{Remove the old bam in the case of the bam realignment}

\item{rstPath}{Path where to save the coverage and quality results. A 'coverage' and 'fastqc' folders will be created.}
}
\description{
Align, infer the coverage and the quality of the bam or fastq files
}
\examples{
\dontrun{

inputFile <- '/imppc/labs/lplab/share/marc/insulinomas/raw/fastq/NET-25_BL_R1.fq.gz'
ref <- '/imppc/labs/lplab/share/marc/refgen/hg38/hg38.fa'
outPathBam <- '/imppc/labs/lplab/share/marc/insulinomas/processed/hg38/bam/bwa'
threads <- 4
removeOpt <- 'no'
rstPath <- '/imppc/labs/lplab/share/marc/insulinomas/rst'
wgsSomaticPreprocessing(inputFile = inputFile,
                        ref = ref,
                        outPathBam = outPathBam,
                        threads = threads,
                        removeOpt = removeOpt,
                        rstPath = rstPath)
}

}
