% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamTofastqPicard.R
\name{bamTofastqPicard}
\alias{bamTofastqPicard}
\title{bamTofastqPicard}
\usage{
bamTofastqPicard(
  bam,
  ref,
  out_path,
  threads,
  gatk4 = "/imppc/labs/lplab/share/bin/gatk-4.1.3.0/gatk",
  sambamba = "sambamba",
  samtools = "samtools"
)
}
\arguments{
\item{bam}{Bam file to carry the analysis.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{gatk4}{Path of GATK4 binary.}

\item{sambamba}{Path of sambamba binary}

\item{samtools}{Path of samtools binary.}
}
\description{
This function converts BAM files to FASTQ files using GATK4 SamToFastq. Type of BAM file (paried or single) is detected
automatically.
}
\examples{
\dontrun{
bamTofastqPicard(bam = 'raw/sample.bam',
             ref = 'ref/hg38.fa',
             out_path = 'rst',
             threads = '2')
}
}
