% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwaAlignment.R
\name{bwaAlignment}
\alias{bwaAlignment}
\title{bwaAlignment}
\usage{
bwaAlignment(input_file, ref, out_path, threads,
  gatk4 = "/imppc/labs/lplab/share/bin/gatk-4.1.3.0/gatk",
  sambamba = "sambamba", bwa = "bwa", samblaster = "samblaster",
  samtools = "samtools")
}
\arguments{
\item{input_file}{Fastq file to align. If paried-end type, `input_file` have to contain mate 1s and different pairs should
be named "_R1" or "_R2". Allowed formats: fastq.gz, fq.gz, fastq, fq or bam.}

\item{ref}{Path for the reference genome to use for the alignment (fasta format) and the corresponding indexes
generated with bwa index and a dictionary index file generated by  CreateSequenceDictionary gatk tool.}

\item{out_path}{Path where the output of the analysis will be saved.}

\item{threads}{Number of threads to use in the analysis.}

\item{gatk4}{Path of GATK4 binary.}

\item{sambamba}{Path of sambamba binary.}

\item{bwa}{Path of bwa binary.}

\item{samblaster}{Path of samblaster binary.}

\item{samtools}{Path of samtools binary.}
}
\description{
Aligns FASTQ files using bwa mem and a defined reference genome. It alignes paired or single end data and BAM files. The
BAM to FASTQ conversion is carried out by Picard's SamToFastq.
}
\examples{
\dontrun{
bwaAlignment(input_file = 'raw/sample_R1.fa',
             ref = 'ref/hg38.fa',
             out_path = 'rst',
             threads = 2)
}
}
